fixed matrix multiplication
values were fixed by hardcoding in a vector sliced in the wrong orientation on other, then dotting it with all the vectors in self and repeating for each slice in other. realized afterwards that it would have been so much better to fix matrix-vector multiplication first and then use that to do matrix-matrix rather than hardcoding in more loops. also want to change makeOrientationRow into makeOrientationMatch(orientation) which would allow me to take out the hardcoding of the orientation as well.