fixed Matrix.rref()
implemented Matrix sort function which rearranges the rows by the index of their leading digit. now rather than popping and appending zero rows during rref checks, the matrix is sorted at the start of every round of checks. swapRows method still exists but is no longer in use since it would be redundant to check if pivots are in the correct order.